gestion suppression fichiers dans la route 'deleteBook'

+++ backend/controllers/stuff.js
@@ -1,5 +1,5 @@
 const Book = require("../models/Book");
-const fs = require("fs");
+const fs = require("fs");//fs = file system
 
 exports.createBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book);
@@ -13,8 +13,6 @@ exports.createBook = (req, res, next) => {
       req.file.filename
     }`,
   });
-  console.log("imageURL //////////////////////////////");
-  console.log(book.imageUrl);
   book
     .save()
     .then(() => res.status(201).json({ message: "Objet enregistré ! " }))
@@ -41,22 +39,36 @@ exports.modifyBook = (req, res, next) => {
   .catch((error)=> {
     res.status(400).json({error})
   })
-  // Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-  //   .then(() => {
-  //     res.status(200).json({ message: "Objet modifié !" });
-  //   })
-  //   .catch((error) => res.status(400).json({ error }));
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+
+  Book.findOne({_id: req.params.id})
+  .then(book => {
+    if(book.userId !== req.auth.userId) {
+      res.status(401).json({message: "Non autorisé"})
+    } else {
+      console.log(book.imageUrl);
+      const filename = book.imageUrl.split("/images/")[1];
+      fs.unlink(`images/${filename}`, () => {
+        Book.deleteOne({_id: req.params.id})
+        .then(() => res.status(200).json({message: "Livre supprimé !"}))
+        .catch(err => res.status(400).json({err}))
+      });
+    }
+  })
+  .catch(error => {
+    res.status(500).json({error})
+  });
+  //Old code
+  // Book.deleteOne({ _id: req.params.id })
+  //   .then(() => res.status(200).json({ message: "Objet supprimé !" }))
+  //   .catch((error) => res.status(400).json({ error }));
 };
 
 exports.getOneBook = (req, res, next) => {
   Book.findOne({ _id: req.params.id })
-    .then((thing) => res.status(200).json(thing))
+    .then((book) => res.status(200).json(book))
     .catch((error) => res.status(400).json({ error }));
 };
